backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncio>=4.0.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.122.0",
    "google>=3.0.0",
    "google-genai>=1.52.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.2.1",
    "pypdf2>=3.0.1",
    "openpyxl>=3.0.0",
    "sqlalchemy>=2.0.44",
    "supabase>=2.24.0",
    "uvicorn>=0.38.0",
]
